pyproject.toml
[tool.poetry]
name = "MyTimer"
version = "0.1.0"
description = ""
authors = ["IvanovIvan1900 <ivanovivan159753@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.10"
django-debugtools = "^2.0"
django-debug-toolbar = "^3.2.4"
django-bootstrap4 = "^21.2"
django-jazzmin = "^2.4.8"
django-environ = "^0.8.1"
django-select2 = "^7.9.0"
django-tempus-dominus = "^5.1.2"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.11.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
commitizen = "^2.20.3"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
pytest-xdist = "^2.5.0"
pytest-lazy-fixture = "^0.6.3"


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
